requirements.txt
# Core ADK and AI
google-adk>=0.1.0
google-genai>=1.0.0

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0

# Data Processing
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Evaluation and Metrics
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.4.0
jupyter>=1.0.0
ipykernel>=6.25.0
